- Re-combined the Adlib slider, only this time with the music volume since I   figured out that the adlib player latched on the music channel.  I really   don't feel like figuring out how to separate them at the moment. - Continued work on the Menu code.  In the process I was able to remove quite   a few #ifdef SPEARs!  This means ECWolf will use a little more memory, but   it is worth it. - Tweaked the volume formula.

+++ src/id_us_1.cpp
@@ -177,7 +177,7 @@ void
 US_PrintUnsigned(longword n)
 {
 	char	buffer[32];
-	sprintf(buffer, "%lu", n);
+	sprintf(buffer, "%lu", static_cast<long unsigned int> (n));
 
 	US_Print(buffer);
 }
